Added some feature functions
In the normal video service side, added a fxn that can clip videos
given the beginning and ending frames (still, there might be some
bugs that I've not yet discovered which might be causing some fast
forwarding effects in clipping).

Also in temporal segmentation part, added an almost copy/paste code of
the whole video segmenting fxn but instead of taking in video path, it
takes in the already decoded features.