Created and filled a 'landmarks' directory
A directory was made for holding all the components that interacted the most
with tasks closely related to landmarks in the images. It includes
finding landmarks, drawing on landmarks, finding velocities and angles
from landmarks.

Also the relevant code path changes were propagated to all other
files, with the addressing done through the new `landmarks` top level directory